LayoutTests/imported/w3c/web-platform-tests/content-security-policy/script-src/script-src-strict_dynamic_double_policy_different_nonce-expected.txt

CONSOLE MESSAGE: The source list for Content Security Policy directive 'script-src' contains an invalid source: ''strict-dynamic''. It will be ignored.
CONSOLE MESSAGE: Refused to load http://localhost:8800/content-security-policy/script-src/simpleSourcedScript.js?unNonced-appendChild because it does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://localhost:8800/content-security-policy/script-src/simpleSourcedScript.js?unNonced-appendChild because it does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://localhost:8800/content-security-policy/script-src/simpleSourcedScript.js?nonced-appendChild because it does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://localhost:8800/content-security-policy/script-src/simpleSourcedScript.js?nonced-appendChild because it does not appear in the script-src directive of the Content Security Policy.
A separate policy with more nonces works correctly with `strict-dynamic` in the script-src directive.


Harness Error (TIMEOUT), message = null

TIMEOUT Unnonced script injected via `appendChild` is not allowed with `strict-dynamic` + a nonce-only double policy. Test timed out
FAIL Script injected via `appendChild` with a correct nonce is allowed with `strict-dynamic` + a nonce-only double policy. assert_unreached: Error should not be triggered. Reached unreachable code
